+++ mdit_py_plugins/admon/index.py
@@ -18,12 +18,10 @@ if TYPE_CHECKING:
     from markdown_it.utils import EnvType, OptionsDict
 
 
-TAGS_RE = re.compile(r'^\s*(?P<tokens>[^"\s](?:[^"]*[^"\s])?)\s+"(?P<title>.*)"\S*$')
-
-
 def _get_multiple_tags(params: str) -> tuple[list[str], str]:
     """Check for multiple tags when the title is double quoted."""
-    match = TAGS_RE.match(params)
+    re_tags = re.compile(r'^\s*(?P<tokens>[^"\s](?:[^"]*[^"\s])?)\s+"(?P<title>.*)"\S*$')
+    match = re_tags.match(params)
     if match:
         tags = match["tokens"].strip().split(" ")
         return [tag.lower() for tag in tags], match["title"]